pyproject.toml
[tool.poetry]
name = "app"
version = "0.0.1"
description = "Webtronics"
authors = ["ssyucfa"]

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "~0.11.3"
python-multipart = "~0.0.5"
pydantic = {extras = ["email"], version = "~1.8.1"}
Jinja2 = "~3.0.1"
SQLAlchemy = "~1.4.3"
alembic = "~1.6.5"
tenacity = "~8.0.1"
httpx = "~0.18.1"
passlib = {extras = ["bcrypt"], version = "^1.7.2"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
cffi = "^1.15.0"
fastapi = "^0.89.0"
